+++ tests/test_cluster/test_silhouette.py
@@ -99,3 +99,53 @@ class SilhouetteVisualizerTests(VisualTestCase):
         Ensure negative silhouette scores are correctly displayed by the visualizer.
         """
         raise NotImplementedError("no negative silhouette example available")
+
+    @pytest.mark.xfail(
+        sys.platform == 'win32', reason="images not close on windows"
+    )
+    def test_colormap_silhouette(self):
+        """
+        Test no exceptions for modifying the colormap in a silhouette visualizer
+        """
+        # Generate a blobs data set
+        X, y = make_blobs(
+            n_samples=1000, n_features=12, centers=8, shuffle=False, random_state=0
+        )
+
+        try:
+            fig = plt.figure()
+            ax = fig.add_subplot()
+
+            visualizer = SilhouetteVisualizer(MiniBatchKMeans(random_state=0), ax=ax, colormap='gnuplot')
+            visualizer.fit(X)
+            visualizer.poof()
+
+            self.assert_images_similar(visualizer)
+        except Exception as e:
+            self.fail("error during silhouette: {}".format(e))
+
+    @pytest.mark.xfail(
+        sys.platform == 'win32', reason="images not close on windows"
+    )
+    def test_colors_silhouette(self):
+        """
+        Test no exceptions for modifying the colors in a silhouette visualizer
+        """
+        # Generate a blobs data set
+        X, y = make_blobs(
+            n_samples=1000, n_features=12, centers=8, shuffle=False, random_state=0
+        )
+
+        try:
+            fig = plt.figure()
+            ax = fig.add_subplot()
+
+            visualizer = SilhouetteVisualizer(MiniBatchKMeans(random_state=0), ax=ax,
+                colors=['red', 'green', 'blue', 'indigo', 'cyan', 'lavender']
+            )
+            visualizer.fit(X)
+            visualizer.poof()
+
+            self.assert_images_similar(visualizer)
+        except Exception as e:
+            self.fail("error during silhouette: {}".format(e))